backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

cdp-sdk

httpx>=0.25.0
web3>=6.11.0
eth-abi>=4.0.0
eth-utils>=2.3.0

langchain>=0.1.4
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2

python-dotenv>=1.0.0
nest-asyncio>=1.6.0

